% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgg-package.R
\docType{package}
\name{pkgg-package}
\alias{pkgg}
\alias{pkgg-package}
\title{pkgg: Just a smol package handling package}
\description{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/onwhenrdy/pkgg}
  \item Report bugs at \url{https://github.com/onwhenrdy/pkgg/issues}
}

}
\author{
\strong{Maintainer}: Dominik Selzer \email{dominik.selzer@uni-saarland.de} (\href{https://orcid.org/0000-0002-4126-0816}{ORCID})

}
\keyword{internal}
